fix: make GitDirectoryEntry internal
